site_project/requirements.txt
Django==4.1.5
asgiref==3.6.0
mysqlclient==2.1.1
pip==22.3.1
setuptools==65.6.3
sqlparse==0.4.3
wheel==0.38.4
django-ckeditor==6.5.1
Pillow==9.4.0